Create version-2.svelte
This version uses some inline functions for example. If all I am going is toggling a Boolean, then I like the inline method, but if there are 2 or more lines of code in the function then I prefer to use an actual function to do it.
Take it for what it is and decide for yourself what one you perfer.